edge_device_sim/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
paho-mqtt==1.6.1
requests==2.31.0
sqlite3
threading
queue
socket
struct
datetime
random
math
logging
json
time
fastapi
uvicorn
websockets 